backend/requirements.txt
uvicorn
gunicorn
python-dotenv
itsdangerous
mypy
loguru
SQLAlchemy
fastapi
pydantic
pydantic[dotenv]
pydantic[email]
annotated-types
anyio
greenlet
idna
psycopg2
sniffio
starlette
typing_extensions